// src/app/models/clients/Physique.ts
import { Adresse } from "../Adresse";

export class Physique {
  id?: any;
  nom?: String;

  prenom?: String;

  mobile?: String;

  fixe?: String;

  email?: String;

  adresse?: Adresse;

  profession?: String;

  lieuDeTravail?: String;

  dateNaissance?: Date;

  photo?: String;

  situationMatrimoniale?: String;
}